+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/install/SequentialInstallation.kt
@@ -3,6 +3,8 @@ package com.atlassian.performance.tools.infrastructure.api.jira.install
 import com.atlassian.performance.tools.infrastructure.api.distribution.ProductDistribution
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraHomeSource
 import com.atlassian.performance.tools.infrastructure.api.jvm.JavaDevelopmentKit
+import com.atlassian.performance.tools.infrastructure.downloadRemotely
+import com.atlassian.performance.tools.infrastructure.installRemotely
 import com.atlassian.performance.tools.ssh.api.SshConnection
 
 class SequentialInstallation(
@@ -15,8 +17,8 @@ class SequentialInstallation(
         ssh: SshConnection,
         server: TcpServer
     ): InstalledJira {
-        val installation = productDistribution.install(ssh, ".")
-        val home = jiraHomeSource.download(ssh)
+        val installation = productDistribution.installRemotely(ssh, ".")
+        val home = jiraHomeSource.downloadRemotely(ssh)
         jdk.install(ssh)
         return InstalledJira(home, installation, jdk, server)
     }